// features/api/model/storeStationRequest.ts
/**
 * Generated by orval v8.0.3 🍺
 * Do not edit manually.
 * Petrol Integrity System API Documentation
 * OpenAPI spec version: 0.0.1
 */

export interface StoreStationRequest {
  /** @maxLength 255 */
  name: string;
  /**
   * @maxLength 255
   * @nullable
   */
  location?: string | null;
  is_active?: boolean;
}