// features/api/model/updateStationRequest.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * Petrol Integrity System API Documentation
 * OpenAPI spec version: 0.0.1
 */

export interface UpdateStationRequest {
  /** @maxLength 255 */
  name?: string;
  /**
   * @maxLength 255
   * @nullable
   */
  location?: string | null;
  is_active?: boolean;
}